added APR tests file
